+++ modules/aws_exocompute/main.tf
@@ -17,6 +17,7 @@ data "polaris_aws_account" "account" {
 # Create an Exocompute configuration.
 resource "polaris_aws_exocompute" "configuration" {
   account_id                = var.cloud_account_id
+  cluster_access            = var.cluster_access
   cluster_security_group_id = var.cluster_security_group_id
   node_security_group_id    = var.node_security_group_id
   region                    = var.region