+++ CHANGELOG.md
@@ -15,3 +15,4 @@
 * Add support for the `CLOUD_DISCOVERY` feature to the `aws_iam_account` module.
 * Add support for private Exocompute to the `aws_exocompute` module.
 * Add support for pod subnets to the `aws_exocompute` module.
+* Add CIDR format validation to the `aws_exocompute` module's `exocompute_vpc` submodule.