+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## v1.2.0
+* Add support for AWS role chaining to the `aws_iam_account` module.
+* Add CIDR format validation to the `aws_exocompute` module's `exocompute_vpc` submodule.
+
 ## v1.1.0
 * Add `gcp_project` Terraform module.
 * Add `gcp_archival_location` Terraform module.
@@ -15,4 +19,3 @@
 * Add support for the `CLOUD_DISCOVERY` feature to the `aws_iam_account` module.
 * Add support for private Exocompute to the `aws_exocompute` module.
 * Add support for pod subnets to the `aws_exocompute` module.
-* Add CIDR format validation to the `aws_exocompute` module's `exocompute_vpc` submodule.